cleanup macro manager
